Move piece to Pieces/.
