+++ osu.Game.Modes.Taiko/osu.Game.Modes.Taiko.csproj
@@ -52,7 +52,7 @@
     <Compile Include="Judgements\TaikoDrumRollTickJudgement.cs" />
     <Compile Include="Judgements\TaikoJudgement.cs" />
     <Compile Include="Judgements\TaikoHitResult.cs" />
-    <Compile Include="Objects\Drawable\CentreHitCirclePiece.cs" />
+    <Compile Include="Objects\Drawable\Pieces\CentreHitCirclePiece.cs" />
     <Compile Include="Objects\Drawable\DrawableCentreHit.cs" />
     <Compile Include="Objects\Drawable\DrawableHit.cs" />
     <Compile Include="Objects\Drawable\DrawableStrongCentreHit.cs" />
